+++ pages/moments-media-ui.js
@@ -15,7 +15,7 @@ import {
   migrateVideoSectionMedia,
   migrateMusicSectionMedia,
   migrateLetterMediaSection
-} from "./moment-media.js?v=216";
+} from "./moment-media.js?v=242";
 import { canFitBytes, formatBytes, storageBytesLimit } from "./moment-plans.js?v=237";
 import { getUiLocale } from "./moments-i18n.js?v=216";
 import { FIELD_PHRASE_EN } from "./moments-i18n-fields.js?v=240";